requirements.txt
nltk
pyenchant
tqdm
pandas_ml
matplotlib
